+++ src/index.ts
@@ -1,5 +1,7 @@
 export * as appRunner from './app-runner'
+export * as rds from './rds'
 export * as amplify from './amplify'
+export * as vercel from './vercel'
 export * as cloudflare from './cloudflare'
 export * as s3 from './s3'
 export * as route53 from './route53'